IDVA3-1037
Add static content to individual statement page.
