+++ locales/cy/individual-statement.json
@@ -1,0 +1,6 @@
+{
+    "individual_statement_title" : "To be translated",
+    "individual_statement_value" : "To be translated",
+    "individual_statement_part1" : "To be translated ",
+    "individual_statement_part2" :  " To be translated."
+}
