+++ locales/en/individual-statement.json
@@ -1,0 +1,6 @@
+{
+    "individual_statement_title" : "Confirm the identity verification statement",
+    "individual_statement_value" : "individual_statement",
+    "individual_statement_part1" : "I confirm that ",
+    "individual_statement_part2" :  " has verified their identity."
+}
