+++ src/constants.ts
@@ -11,7 +11,7 @@ export const Urls = {
     PSC_TYPE: "/psc-type",
     SKELETON_THREE: "/skeleton_three",
     PERSONAL_CODE: "/individual/personal-code",
-    SKELETON_FIVE: "/skeleton_five",
+    INDIVIDUAL_STATEMENT: "/individual/statement",
     PSC_VERIFIED: "/psc-verified",
     HEALTHCHECK: "/healthcheck"
 } as const;
@@ -25,7 +25,7 @@ export const PrefixedUrls = {
     PSC_TYPE: servicePathPrefix + Urls.PSC_TYPE,
     SKELETON_THREE: servicePathPrefix + Urls.SKELETON_THREE,
     PERSONAL_CODE: servicePathPrefix + Urls.PERSONAL_CODE,
-    SKELETON_FIVE: servicePathPrefix + Urls.SKELETON_FIVE,
+    INDIVIDUAL_STATEMENT: servicePathPrefix + Urls.INDIVIDUAL_STATEMENT,
     PSC_VERIFIED: servicePathPrefix + Urls.PSC_VERIFIED,
     COOKIES: "/help/cookies"
 } as const;
