+++ src/router.dispatch.ts
@@ -1,6 +1,6 @@
 // Do Router dispatch here, i.e. map incoming routes to appropriate router
 import { Application, Request, Response, Router } from "express";
-import { StartRouter, CompanyNumberRouter, ConfirmCompanyRouter, PscTypeRouter, SkeletonThreeRouter, PersonalCodeRouter, SkeletonFiveRouter, PscVerifiedRouter } from "./routers/__utils";
+import { StartRouter, CompanyNumberRouter, ConfirmCompanyRouter, PscTypeRouter, SkeletonThreeRouter, PersonalCodeRouter, PscVerifiedRouter, IndividualStatementRouter } from "./routers/__utils";
 import { Urls, servicePathPrefix } from "./constants";
 
 const routerDispatch = (app: Application) => {
@@ -16,7 +16,7 @@ const routerDispatch = (app: Application) => {
     router.use(Urls.PSC_TYPE, PscTypeRouter);
     router.use(Urls.SKELETON_THREE, SkeletonThreeRouter);
     router.use(Urls.PERSONAL_CODE, PersonalCodeRouter);
-    router.use(Urls.SKELETON_FIVE, SkeletonFiveRouter);
+    router.use(Urls.INDIVIDUAL_STATEMENT, IndividualStatementRouter);
     router.use(Urls.PSC_VERIFIED, PscVerifiedRouter);
     router.use("*", (req: Request, res: Response) => {
         res.status(404).render("partials/error_400");
