+++ src/routers/__utils/index.ts
@@ -4,6 +4,6 @@ import ConfirmCompanyRouter from "./../confirmCompanyRouter";
 import PscTypeRouter from "./../psc.type.router";
 import SkeletonThreeRouter from "./../skeleton.three.router";
 import PersonalCodeRouter from "../personalCode.router";
-import SkeletonFiveRouter from "./../skeleton.five.router";
+import IndividualStatementRouter from "../individual.statement.router";
 import PscVerifiedRouter from "./../pscVerifiedRouter";
-export { StartRouter, CompanyNumberRouter, ConfirmCompanyRouter, PscTypeRouter, SkeletonThreeRouter, PersonalCodeRouter, SkeletonFiveRouter, PscVerifiedRouter };
+export { StartRouter, CompanyNumberRouter, ConfirmCompanyRouter, PscTypeRouter, SkeletonThreeRouter, PersonalCodeRouter, IndividualStatementRouter, PscVerifiedRouter };
