+++ src/routers/individual.statement.router.ts
@@ -1,11 +1,11 @@
 import { NextFunction, Request, Response, Router } from "express";
-import { SkeletonFiveHandler } from "./handlers/skeleton_five";
+import { IndividualStatementHandler } from "./handlers/individual_statement/individual_statement";
 import { handleExceptions } from "../utils/async.handler";
 import logger from "../lib/Logger";
 const router: Router = Router();
 
 router.get("/", handleExceptions(async (req: Request, res: Response) => {
-    const handler = new SkeletonFiveHandler();
+    const handler = new IndividualStatementHandler();
     const { templatePath, viewData } = handler.executeGet(req, res);
     res.render(templatePath, viewData);
 }));
