+++ src/views/router_views/personal_code/personal_code.njk
@@ -28,7 +28,7 @@
       }) }}
       {{ govukButton({
           text: i18n.global_save_continue_button,
-          href: Urls.SKELETON_FIVE,
+          href: Urls.INDIVIDUAL_STATEMENT,
           classes: "govuk-!-margin-top-2 govuk-!-margin-bottom-8"
       }) }}
     </div>
