--- src/views/router_views/skeleton_five/skeleton_five.njk
@@ -1,18 +1,0 @@
-{% extends "layouts/default.njk" %}
-
-{% from "govuk/components/button/macro.njk" import govukButton %}
-
-{% block main_content %}
-
-  <h1 class="govuk-heading-xl" id="presenter-account--title">
-      {{title}}
-  </h1>
-
-      {{ govukButton({
-          text: "Skeleton Six",
-          href: Urls.PSC_VERIFIED,
-          classes: "govuk-button--start govuk-!-margin-top-2 govuk-!-margin-bottom-8",
-          isStartButton: true
-      }) }}
-
-{% endblock %}
